support parameters in stored queries not only in filters
Currently, parameters are only supported in filter expressions in a stored query. Parameters are also useful in other members of a query expression, e.g. the CRS information, the limit value or the profiles to apply.

This commit supports the use of parameters in all members of the query expression of a stored query, except "id", "title", "description", and "offset".

A new immutable `StoredQueryExpression` is introduced that supports parameter objects in the values of the query expression of a stored query. These instances are what is stored as values. The `QueryExpression` is simplified to remove all support for parameters.

Example of the use of parameters in "crs" and "profiles":

```
    ...,
    "crs": {
        "$parameter": {
            "crs": {
                "title": "Koordinatenreferenzsystem",
                "description": "Koordinatenreferenzsystem in dem Geometrien zurückgegeben werden sollen.",
                "type": "string",
                "enum": [
                    "http://www.opengis.net/def/crs/EPSG/0/25832" ,
                    "http://www.opengis.net/def/crs/EPSG/0/4258"  ,
                    "http://www.opengis.net/def/crs/OGC/1.3/CRS84"
                ],
                "default": "rhttp://www.opengis.net/def/crs/EPSG/0/25832"
            }
        }
    },
    "profiles": [
        {
            "$parameter": {
                "profil-rel": {
                    "title": "Anzuwendenes Profil für Relationen",
                    "description": "Das Profil für die Darstellung von Objektreferenzen, das angewendet werden soll. `rel-as-key` bedeutet, dass die Objektreferenzen als Fremdschlüssel des referenzierten Objekts dargestellt werden; bei `rel-as-uri` wird die HTTP-URI des referenzierten Objekts verwendet; bei `rel-as-link` wird ein Link-Objekt verwendet.",
                    "type": "string",
                    "enum": ["rel-as-key", "rel-as-uri", "rel-as-link"],
                    "default": "rel-as-uri"
                }
            }
        },
        {
            "$parameter": {
                "profil-val": {
                    "title": "Anzuwendenes Profil für Codelisten",
                    "description": "Das Profil für die Darstellung von Werten aus Codelisten, das angewendet werden soll. Bei `val-as-code` werden die Werte als  dargestellt; bei `val-as-title` werden die Werte als lesbare Texte dargestellt.",
                    "type": "string",
                    "enum": ["val-as-code", "val-as-title"],
                    "default": "val-as-code"
                }
            }
        }
    ],
    ...
```
