+++ ogcapi-draft/ogcapi-features-search/src/main/java/de/ii/ogcapi/features/search/app/StoredQueryExpressionFactory.java
@@ -8,17 +8,17 @@
 package de.ii.ogcapi.features.search.app;
 
 import com.github.azahnen.dagger.annotations.AutoBind;
-import de.ii.ogcapi.features.search.domain.QueryExpression;
+import de.ii.ogcapi.features.search.domain.StoredQueryExpression;
 import de.ii.xtraplatform.values.domain.ValueFactoryAuto;
 import javax.inject.Inject;
 import javax.inject.Singleton;
 
 @Singleton
 @AutoBind
-public class QueryExpressionFactory extends ValueFactoryAuto {
+public class StoredQueryExpressionFactory extends ValueFactoryAuto {
 
   @Inject
-  QueryExpressionFactory() {
-    super(QueryExpression.class);
+  StoredQueryExpressionFactory() {
+    super(StoredQueryExpression.class);
   }
 }
