+++ ogcapi-draft/ogcapi-features-search/src/main/java/de/ii/ogcapi/features/search/domain/StoredQueryFormat.java
@@ -26,5 +26,6 @@ public interface StoredQueryFormat extends GenericFormatExtension {
    */
   String getFileExtension();
 
-  Object getEntity(QueryExpression query, OgcApiDataV2 apiData, ApiRequestContext requestContext);
+  Object getEntity(
+      StoredQueryExpression query, OgcApiDataV2 apiData, ApiRequestContext requestContext);
 }
